Adds sqlite support
Refactors db.js into separate modules
