+++ package.json
@@ -1,42 +1,40 @@
 {
-	"version": "1.0.3",
-	"name": "mergeatron",
-	"description": "A helpful PR monitor that runs jenkins builds when Pull Requests are created and updated",
-	"author": "SNAP Interactive, Inc.",
-	"contributors": [
-		"Steven Surowiec <steven.surowiec@gmail.com>",
-		"Mike Sherov <mike.sherov@gmail.com>"
-	],
-
-	"licenses": [{
-		"type": "Apache-2.0",
-		"url": "http://www.apache.org/licenses/LICENSE-2.0"
-	}],
-
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/SnapInteractive/mergeatron"
-	},
-
-	"dependencies": {
-		"github": ">=0.1.5",
-		"mongodb": "1.3.19",
-		"mongojs": ">=0.4.3",
-		"mysql": "2.0.0-alpha4",
-		"request": ">=2.9.203",
-		"node-uuid": ">=1.3.3",
-		"async": ">=0.1.22",
-		"winston": ">=0.6.2",
-		"xml2js": ">=0.2.2",
-		"range_check": "*"
-	},
-
-	"devDependencies": {
-		"grunt": "0.4.0a",
-		"grunt-contrib-nodeunit": "~0.1.1",
-		"grunt-contrib-jshint": "~0.1.0",
-		"grunt-contrib-watch": "~0.1.4"
-	},
-
-	"main": "./mergeatron.js"
+  "version": "1.0.3",
+  "name": "mergeatron",
+  "description": "A helpful PR monitor that runs jenkins builds when Pull Requests are created and updated",
+  "author": "SNAP Interactive, Inc.",
+  "contributors": [
+    "Steven Surowiec <steven.surowiec@gmail.com>",
+    "Mike Sherov <mike.sherov@gmail.com>"
+  ],
+  "licenses": [
+    {
+      "type": "Apache-2.0",
+      "url": "http://www.apache.org/licenses/LICENSE-2.0"
+    }
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/SnapInteractive/mergeatron"
+  },
+  "dependencies": {
+    "async": ">=0.1.22",
+    "github": ">=0.1.5",
+    "mongodb": "1.3.19",
+    "mongojs": ">=0.4.3",
+    "mysql": "2.0.0-alpha4",
+    "node-uuid": ">=1.3.3",
+    "range_check": "*",
+    "request": ">=2.9.203",
+    "sqlite3": "^3.0.8",
+    "winston": ">=0.6.2",
+    "xml2js": ">=0.2.2"
+  },
+  "devDependencies": {
+    "grunt": "0.4.0a",
+    "grunt-contrib-nodeunit": "~0.1.1",
+    "grunt-contrib-jshint": "~0.1.0",
+    "grunt-contrib-watch": "~0.1.4"
+  },
+  "main": "./mergeatron.js"
 }
